fix(NANO-115): move history mode toggle from Settings to Launcher page
Toggle was placed in the Settings/Providers read-only card instead of the
Launcher's LLM Configuration card where the user specified. Moved to
LLMSection component, visible below the Local/Cloud tabs regardless of
provider type. Reverted provider-display.tsx to its original state.
