+++ gui/src/components/launcher/llm-section.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useState, useCallback, useEffect, useRef } from "react";
-import { Brain, Cloud, HardDrive } from "lucide-react";
+import { Brain, Cloud, GitBranch, HardDrive } from "lucide-react";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Label } from "@/components/ui/label";
 import { Input } from "@/components/ui/input";
@@ -17,7 +17,8 @@ import {
   ModelCombobox,
   type ModelOption,
 } from "@/components/ui/model-combobox";
-import { useLauncherStore } from "@/lib/stores";
+import { useLauncherStore, useSettingsStore } from "@/lib/stores";
+import { getSocket } from "@/lib/socket";
 import { CLOUD_PROVIDERS, CLOUD_PROVIDER_IDS, type CloudProvider } from "@/lib/constants/cloud-providers";
 
 interface FieldRowProps {
@@ -390,6 +391,26 @@ function CloudLLMFields() {
 
 export function LLMSection() {
   const { llmProviderType, setLLMProviderType } = useLauncherStore();
+  const { generationConfig, setGenerationConfig } = useSettingsStore();
+  const socket = getSocket();
+  const debounceRef = useRef<NodeJS.Timeout | null>(null);
+
+  useEffect(() => {
+    return () => {
+      if (debounceRef.current) clearTimeout(debounceRef.current);
+    };
+  }, []);
+
+  const handleHistoryModeChange = useCallback(
+    (mode: "auto" | "splice" | "flatten") => {
+      setGenerationConfig({ ...generationConfig, force_role_history: mode });
+      if (debounceRef.current) clearTimeout(debounceRef.current);
+      debounceRef.current = setTimeout(() => {
+        socket.emit("set_generation_params", { force_role_history: mode });
+      }, 300);
+    },
+    [generationConfig, socket, setGenerationConfig]
+  );
 
   return (
     <Card>
@@ -423,6 +444,31 @@ export function LLMSection() {
             <CloudLLMFields />
           </TabsContent>
         </Tabs>
+
+        <div className="border-t border-border pt-4 mt-4">
+          <p className="text-xs font-medium text-muted-foreground mb-3 flex items-center gap-2">
+            <GitBranch className="h-3 w-3" />
+            History Mode
+          </p>
+          <div className="flex gap-1 rounded-md bg-muted p-1">
+            {(["auto", "splice", "flatten"] as const).map((mode) => (
+              <button
+                key={mode}
+                onClick={() => handleHistoryModeChange(mode)}
+                className={`flex-1 text-xs py-1.5 px-2 rounded transition-colors ${
+                  generationConfig.force_role_history === mode
+                    ? "bg-background text-foreground shadow-sm"
+                    : "text-muted-foreground hover:text-foreground"
+                }`}
+              >
+                {mode === "auto" ? "Auto" : mode === "splice" ? "Splice" : "Flatten"}
+              </button>
+            ))}
+          </div>
+          <p className="text-xs text-muted-foreground mt-2">
+            Auto defers to provider capability. Splice sends real role-array turns (required for Gemma). Flatten embeds history as text in the system prompt.
+          </p>
+        </div>
       </CardContent>
     </Card>
   );
